Move termination_handler.py from streams
