+++ devices/webcam/pose_vis/termination_handler.py
